+++ yatpama.c
@@ -4,6 +4,8 @@
 #include <string.h>
 #include <stdlib.h>
 #include <regex.h>
+#include <sys/param.h>
+#include <sys/time.h>
 
 #include "lib/aes.h"
 #include "lib/crypto.h"
@@ -12,6 +14,8 @@
 #include "lib/utilities.h"
 
 #include "yatpama.h"
+
+#define EXEC_VERSION "v1.3.0" // La version de l'exécutable
 
 /*
  * Interface principale
@@ -22,14 +26,15 @@ char prompt() {
     int again;
 
     do {
-        printf("\n----------------------------------------------------------");
-        printf("\n        yatpama : Yet Another Tiny Password Manager       ");
-        printf("\n----------------------------------------------------------");
-        printf("\n k password | p print | s search | a add | d del | q quit ");
-        printf("\n----------------------------------------------------------");
+        printf("\n---------------------------------------------------------------------------");
+        printf("\n                yatpama : Yet Another Tiny Password Manager                ");
+        printf("\n---------------------------------------------------------------------------");
+        printf("\n k pwd | p print | s search | a add | d del | e export | i import | q quit ");
+        printf("\n---------------------------------------------------------------------------");
         printf("\nChoose a command: ");
         cmd = getchar();
-        again = cmd != 'k' && cmd != 'p' && cmd != 'a' && cmd != 'q' && cmd != 's' && cmd != 'd'; 
+        again = cmd != 'k' && cmd != 'p' && cmd != 'a' && cmd != 'q' &&
+                cmd != 's' && cmd != 'd' && cmd != 'e' && cmd != 'i'; 
     } while (again);
 
     getchar(); // Enlever la touche 'Enter' du buffer du clavier
@@ -38,14 +43,41 @@ char prompt() {
 }
 
 /*
- * Saisir le mot de passe pour générer une clé principale pour chiffrer / déchiffrer
- * La fonction génère la clé (paramètre)
+ * Calculer la valeur de hachage du fichier exécuté
+ * On prends le chemin absolu pour accéder au fichier exécuté
+ * Paramètre n°1 : le nom du fichier de la ligne de commande
+ * Paramètre n°2 : la valeur de hachage calculée
  */
-void do_command_key(uint8_t key[]) {
+void get_hash_executable(char* argv0, BYTE hash[]) {
+    // Mise à zéro du hash
+    memset(hash, 0, 32);
+
+    // Récupérer le chemin absolu complet de l'executable
+    char path[MAXPATHLEN*2];
+    *path = '\0';
+    getAbsolutePath("yatpama", argv0, path, sizeof(path));
+
+    if (*path != '\0') {
+        char * canonical_path = realpath(path, NULL);
+        printf("\nReference used: %s\n", canonical_path);
+        compute_hash_executable(path, hash);
+        free(canonical_path);
+    } else {
+        fprintf(stderr, "Impossible to get access to the executable file!\n");
+        exit(1);
+    }
+}
+
+/*
+ * Saisir le mot de passe pour générer une clé principale pour chiffrer / déchiffrer
+ * La fonction génère la clé (2ème paramètre) de 32 octets soit 256 bits
+ */
+void do_command_key(char * argv0, uint8_t key[]) {
     uint8_t msecret[256];
     char enter;
     int nbChar;
 
+    memset(msecret, 0, 256);
     printf("Master password: ");
     nbChar = scanf("%s", msecret);
 
@@ -54,10 +86,69 @@ void do_command_key(uint8_t key[]) {
     	pwdConformity(msecret, PWD_SIZE); // Contrôle de la conformité du mot de passe
         pwdtokey(msecret, 256, key); // Génére une clé à partir du mot de passe
     	memset(msecret, 0, 256); // On oublie le mot de passe
+
+        // Génère la clé finale = key_from_pwd xor hash_from_executable
+        BYTE hash[AES_KEYLEN];
+        get_hash_executable(argv0, hash);
+        xor_table(key, hash, AES_KEYLEN);
     } else {
         fprintf(stderr, "Impossible to read the password!\n");
         exit(1);
     }
+}
+
+/*
+ * Chiffrement d'une entrée
+ * Paramètre 1 : la clé de chiffrement
+ * Paramètre 2 : l'entrée contenant les deux chaines claires puis chiffrées
+ */
+void cypher_data(uint8_t key[], Entry * pentry) {
+    // Construction de text en concaténant les deux chaines
+    // text == information | secret
+    uint8_t text[MAX_SIZE * 2];
+    memset(text, 0, sizeof text); // Mise à zéro de la zone mémoire text
+    concat(text, pentry->information);
+    concat(text, pentry->secret);
+
+    // Calcul du HMAC sur la chaine text
+    hmac_sha256(text, sizeof text, key, AES_KEYLEN, pentry->hash);
+
+    struct AES_ctx ctx;
+
+    // Générer IV pour chiffrer information
+    rng(pentry->iv_info, AES_BLOCKLEN);
+
+    // Chiffrer information
+    AES_init_ctx_iv(&ctx, key, pentry->iv_info);
+    AES_CBC_encrypt_buffer(&ctx, pentry->information, sizeof pentry->information);
+
+    // Générer IV pour secret
+    rng(pentry->iv_sec, sizeof pentry->iv_sec);
+
+    // Chiffrer secret
+    AES_init_ctx_iv(&ctx, key, pentry->iv_sec);
+    AES_CBC_encrypt_buffer(&ctx, pentry->secret, sizeof pentry->secret);
+}
+
+/*
+ * Déchiffrement en mémoire d'une entrée 
+ * Paramètre 1 : la clé de déchiffrement
+ * Paramètre 2 : l'entrée contenant les 2 données chiffrées
+ * Paramètre 3 : pointeur sur l'information déchiffrée
+ * Paramètre 4 : pointeur sur le secret déchiffré
+ */
+void uncypher_data(uint8_t key[], Entry * pentry, uint8_t * pinformation, uint8_t * psecret) {
+    struct AES_ctx ctx;
+
+    // Déchiffrement de information en mémoire
+    memcpy(pinformation, pentry->information, MAX_SIZE);
+    AES_init_ctx_iv(&ctx, key, pentry->iv_info);
+    AES_CBC_decrypt_buffer(&ctx, pinformation, MAX_SIZE);
+
+    // Déchiffrement de secret en mémoire
+    memcpy(psecret, pentry->secret, MAX_SIZE);
+    AES_init_ctx_iv(&ctx, key, pentry->iv_sec);
+    AES_CBC_decrypt_buffer(&ctx, psecret, MAX_SIZE);
 }
 
 /*
@@ -77,12 +168,8 @@ void search_and_print(uint8_t key[], DLList list, char* pattern, int pos) {
         int nbInfo = 0;
         int nbInfoMatch = 0;
 
-        Entry * pentry; // Pointeur sur l'enregistrement
-
         uint8_t information[MAX_SIZE];
         uint8_t secret[MAX_SIZE];
-
-        struct AES_ctx ctx;
 
         // Compilation de l'expression régulière à partir du motif
         regex_t reg;
@@ -96,18 +183,8 @@ void search_and_print(uint8_t key[], DLList list, char* pattern, int pos) {
         }
 
         do {
-            // Obtenir l'enregistrement courant de la liste
-            pentry = list->entry;
-
-            // Déchiffrement de information en mémoire
-            memcpy(information, pentry->information, MAX_SIZE);
-            AES_init_ctx_iv(&ctx, key, pentry->iv_info);
-            AES_CBC_decrypt_buffer(&ctx, information, sizeof information);
-
-            // Déchiffrement de secret en mémoire
-            memcpy(secret, pentry->secret, MAX_SIZE);
-            AES_init_ctx_iv(&ctx, key, pentry->iv_sec);
-            AES_CBC_decrypt_buffer(&ctx, secret, sizeof secret);
+            // Déchiffrement en mémoire
+            uncypher_data(key, list->entry, information, secret);
 
             // Pattern matching
             int match1 = 0;
@@ -189,34 +266,13 @@ DLList do_command_add(uint8_t key[], DLList list) {
     printf("Secret: ");
     getsl((char*)pentry->secret, MAX_SIZE);
 
-    // Construction de text en concaténant les deux chaines
-    // text == information | secret
-    uint8_t text[MAX_SIZE * 2];
-    memset(text, 0, sizeof text); // Mise à zéro de la zone mémoire text
-    concat(text, pentry->information);
-    concat(text, pentry->secret);
+    // Chiffrement de l'entrée
+    cypher_data(key, pentry);
 
-    struct AES_ctx ctx;
-
-    // Générer IV pour chiffrer information
-    rng(pentry->iv_info, AES_BLOCKLEN);
-
-    // Chiffrer information
-    AES_init_ctx_iv(&ctx, key, pentry->iv_info);
-    AES_CBC_encrypt_buffer(&ctx, pentry->information, sizeof pentry->information);
-
-    // Générer IV pour secret
-    rng(pentry->iv_sec, sizeof pentry->iv_sec);
-
-    // Chiffrer secret
-    AES_init_ctx_iv(&ctx, key, pentry->iv_sec);
-    AES_CBC_encrypt_buffer(&ctx, pentry->secret, sizeof pentry->secret);
-
-    // Calcul du HMAC sur la chaine text
-    hmac_sha256(text, sizeof text, key, AES_KEYLEN, pentry->hash);
-
-    // Ajouter de l'entrée dans la liste
+    // Ajout de l'entrée dans la liste
     list = addAtLast_DLList(list, pentry);
+
+    printf("\nOne entry added\n");
 
     return list;
 }
@@ -234,16 +290,20 @@ DLList do_command_add(uint8_t key[], DLList list) {
 DLList do_command_delete(uint8_t key[], DLList list) {
     const int nbChiffre = 4; // 9999 maximum d'entrées
     int erreur; // Drapeau indicateur d'une erreur
+    char * pret; // Pointeur sur valeur retournée par fgets
 
     char cNbEntry[nbChiffre + 1]; // Numéro de l'entrée en chaine
     int nbEntry; // Numéro de l'entrée en entier
     
     // Obtenir et tester le numéro de l'entrée à supprimer
     printf("Give entry number: ");
-    fgets(cNbEntry, nbChiffre + 1, stdin);
+    pret = fgets(cNbEntry, nbChiffre + 1, stdin);
+    erreur = pret == NULL;
 
-    nbEntry = atoi(cNbEntry);
-    erreur = nbEntry <= 0 || nbEntry > size_DLList(list);
+    if (!erreur) {
+        nbEntry = atoi(cNbEntry);
+        erreur = nbEntry <= 0 || nbEntry > size_DLList(list);
+    }
 
     if (!erreur) {
         char response[] = "n";
@@ -253,10 +313,10 @@ DLList do_command_delete(uint8_t key[], DLList list) {
 
         // Demander confirmation
         printf("\nPlease, confirm you want delete this entry [y/n]: ");
-        fgets(response, 2, stdin);
+        pret = fgets(response, 2, stdin);
 
         // Confirmation positive : supppresion de l'entrée de la liste
-        if (response[0] == 'y') {
+        if (pret != NULL && response[0] == 'y') {
             list = del_Element_DLList(list, nbEntry);
             nbEntry = size_DLList(list);
             printf("Confirmation: one entry deleted, %d entries left.\n", nbEntry);
@@ -268,8 +328,74 @@ DLList do_command_delete(uint8_t key[], DLList list) {
         printf("\nThis entry number does not exist\n");
     }
 
-    fpurge(stdin);
+    fflush(stdin);
     return list;
+}
+
+/*
+ * Chargement et contrôle de l'enregistrement spécial de contrôle de version
+ * Paramètre 1 : le numéro de fichier data ouvert
+ * Paramètre 2 : la clé principale pour le contrôle du hmac
+ */
+void load_special_entry(int fp, uint8_t key[]) {
+    int nblus;
+    int erreur = 0;
+    Entry entry;
+
+    memset(&entry, 0, sizeof entry);
+
+    // ---------------------------
+    // Lecture de l'enregistrement
+    nblus = read(fp, &entry, sizeof entry);
+    erreur = nblus != sizeof entry;
+
+    if (erreur) {
+        fprintf(stderr, "Impossible to read the special entry from data file!\n");
+        close(fp);
+        exit(1);
+    }
+
+    // ----------------
+    // Contrôle du hash
+
+    uint8_t hash2[HASH_SIZE];  // Hash de contrôle
+    uint8_t text[MAX_SIZE * 2];
+    memset(text, 0, sizeof text); // Mise à zéro de la zone mémoire text
+
+    // Construction de text en concaténant les deux chaines
+    // text == information | secret
+    concat(text, entry.information);
+    concat(text, entry.secret);
+    
+    // Calcul du hash de contrôle
+    hmac_sha256(text, sizeof text, key, AES_KEYLEN, hash2);
+
+    // Comparaison du hash
+    int erreurHash = 0 != memcmp((const void *)&entry.hash, (const void *)&hash2, sizeof hash2);
+
+    // ------------------------------------
+    // Prise en compte du numéro de version
+    int erreurVersion = 0 != memcmp((const void*)EXEC_VERSION, (const void *)& entry.information, sizeof EXEC_VERSION);
+
+    // -------------------------
+    // Gestion des cas d'erreurs
+    if (erreurHash && erreurVersion) {
+        printf("\nA previous version (%s) has been detected for the data file", entry.information);
+        printf("\nThe current version used is %s", EXEC_VERSION);
+        printf("\nSee the procedure at https://github.com/thethythy/yatpama to retrieve data safely\n\n");
+        close(fp);
+        exit(1);
+    }
+    else if (erreurHash && !erreurVersion) {
+        fprintf(stderr, "\nWrong password or data file has been corrupted!\n");
+        close(fp);
+        exit(1);
+    }
+
+    // -------------------------------------
+    // Bouclier anti attaque par force brute
+    // TODO
+
 }
 
 /*
@@ -284,6 +410,9 @@ DLList load_data(uint8_t key[], const char *file_name) {
     fp = open(file_name, O_RDONLY);
 
     if (fp != -1) {
+
+        // Lecture et contrôle de l'enregistrement spécial
+        load_special_entry(fp, key);
 
         int nblus;
         int erreur = 0;
@@ -327,7 +456,7 @@ DLList load_data(uint8_t key[], const char *file_name) {
 
                 // Comparaison de hash avec hash2 (fonction compare)
                 if (-1 == compare(pentry->hash, sizeof pentry->hash, hash2, sizeof hash2)) {
-                    fprintf(stderr, "\nWrong password or wrong file or data has been corrupted!\n");
+                    fprintf(stderr, "\nWrong password or data file has been corrupted!\n");
                     close(fp);
                     exit(1);
                 }
@@ -417,6 +546,55 @@ void backup_data(const char *file_name) {
 }
 
 /**
+ * Construction de l'enregistrement spécial de contrôle de version
+ * Paramètre n°1 : fp est le numéro du fichier data ouvert en écriture
+ * Paramètre n°2 : la clé pour le calcul du hmac
+ */
+void save_special_entry(int fp, uint8_t key[]) {
+    ssize_t nbBytes;
+    int erreur = 0; 
+    Entry entry;
+    
+    // Mise à zéro
+    memset(&entry, 0, sizeof entry);
+
+    // Numéro de version
+    memcpy((void *)entry.information, (const void *)EXEC_VERSION, sizeof EXEC_VERSION);
+
+    // Date en secondes
+    struct timeval time;
+    gettimeofday(&time, NULL);
+    erreur = sizeof time.tv_sec >= sizeof entry.secret;
+
+    if (!erreur) {
+        memcpy((void *)entry.secret, (const void*)& (time.tv_sec), sizeof time.tv_sec);
+    }
+
+    // Calcul d'un HMAC
+    if (!erreur) {
+        // Construction de text en concaténant les deux chaines
+        // text == information | secret
+        uint8_t text[MAX_SIZE * 2];
+        memset(text, 0, sizeof text); // Mise à zéro de la zone mémoire text
+        concat(text, entry.information);
+        concat(text, entry.secret);
+        hmac_sha256(text, sizeof text, key, AES_KEYLEN, entry.hash);
+    }
+
+    // Ecriture de l'enregistrement spéciale
+    if (!erreur) {
+        nbBytes = write(fp, &entry, sizeof entry);
+        erreur = nbBytes != sizeof entry;
+    }
+
+    if (erreur) {
+        fprintf(stderr, "Impossible to create the special entry in data file!\n");
+        close(fp);
+        exit(1);
+    }
+}
+
+/**
  * Enregistre les données chiffrées dans le fichier
  * La fonction connait la liste des données (premier paramètre)
  * La fonction connait le nom complet du fichier de données (second paramètre)
@@ -424,7 +602,7 @@ void backup_data(const char *file_name) {
  * On consdidère que la liste n'est pas vide au départ !
  * Si ce fichier existe déjà, on fait une copie de sauvegarde avant de l'écraser
  */
-void save_data(DLList list, const char *file_name){
+void save_data(DLList list, const char *file_name, uint8_t key[]){
     int fp;
     
     // Test de l'existence du fichier
@@ -442,18 +620,24 @@ void save_data(DLList list, const char *file_name){
     // Si un fichier existe on écrit dedans
     if (fp != -1) {
 
+        // Construction de l'enregistrement spéciale
+        save_special_entry(fp, key);
+
+        // Ecriture des entrées de la liste
         do {
-            
+
             ssize_t nbBytes;
-            int erreur = 0;
+            int erreur = 0;          
             Entry * pentry;
 
-            // On prend l'enrée en tête de liste
-            pentry = list->entry;
+            if (!isEmpty_DLList(list)) {
+                // On prend l'entrée en tête de liste
+                pentry = list->entry;
 
-            // Ecriture de la structure dans le fichier
-            nbBytes = write(fp, pentry, sizeof *pentry);
-            erreur = nbBytes != sizeof *pentry;
+                // Ecriture de la structure dans le fichier
+                nbBytes = write(fp, pentry, sizeof *pentry);
+                erreur = nbBytes != sizeof *pentry;
+            }
 
             if (erreur) {
                 fprintf(stderr, "Impossible to write in data file!\n");
@@ -476,6 +660,155 @@ void save_data(DLList list, const char *file_name){
     }
 }
 
+/*
+ * Exporte les données en clair dans un fichier texte
+ * La fonction connait la clé de déchiffrement (premier paramètre)
+ * La fonction connait la liste des données chiffrées (second paramètre)
+ * La fonction connait le nom du fichier d'exportation (troisième paramètre)
+ */
+void do_command_export(uint8_t key[], DLList list, const char *file_export) {
+    if (!isEmpty_DLList(list)) {
+
+        int fp;
+
+        // Création ou ouverture en mode écrasement
+        if (access(file_export, F_OK) == -1)
+            fp = creat(file_export, 0600);
+        else
+            fp = open(file_export, O_WRONLY | O_TRUNC);
+
+        if (fp != -1) {
+
+            int nbEntries = 0;
+
+            uint8_t information[MAX_SIZE];
+            uint8_t secret[MAX_SIZE];
+
+            char * fin; // Position de la fin de chaine
+            int nbBytes, nbWrote, error;
+
+            do {
+                // Déchiffrement en mémoire
+                uncypher_data(key, list->entry, information, secret);
+
+                // Ecriture dans le fichier d'exportation de "information"
+                fin = index((const char *)information, '\0');
+                nbBytes = fin - (char *)information;
+                nbWrote = write(fp, information, nbBytes);
+                error = nbBytes != nbWrote;
+
+                if (!error) {
+                    nbWrote = write(fp, "\n", 1);
+                    error = 1 != nbWrote;
+                }
+
+                // Ecriture dans le fichier d'exportation de "secret"
+                if (!error) {
+                    fin = index((const char *)secret, '\0');
+                    nbBytes = fin - (char *)secret;
+                    nbWrote = write(fp, secret, nbBytes);
+                    error = nbBytes != nbWrote;
+                }
+
+                if (!error) {
+                    nbWrote = write(fp, "\n", 1);
+                    error = 1 != nbWrote;
+                }
+                
+                if (error) {
+                    fprintf(stderr, "\nImpossible to write to the exportation file (%s)\n", file_export);
+                    close(fp);
+                    exit(1);
+                }
+
+                nbEntries++;
+
+                // Mise à zéro des zones mémoires utilisées
+                memset(information, 0, sizeof information);
+                memset(secret, 0, sizeof secret);
+
+                list = next_DLList(list); // Noeud suivant
+
+            } while(!isEmpty_DLList(list));
+
+            printf("\n%d entries has been exported in %s\n", nbEntries, file_export);
+            close(fp);
+
+        } else {
+            fprintf(stderr, "\nImpossible to create or open to the exportation file!\n");
+            exit(1);
+        }
+
+    } else
+        printf("\nThere is no entry yet!\n");
+}
+
+/*
+ * Importe des données depuis un fichier texte
+ * La fonction connait la clé de chiffrement (premier paramètre)
+ * La fonction connait la liste des données chiffrées (second paramètre)
+ * La liste (éventuellement modifiée) est retournée
+ */
+DLList do_command_import(uint8_t key[], DLList list) {
+    char file_import[MAXPATHLEN];
+    char * pret; // Pointeur sur la valeur de retour de fgets
+    FILE * pf;
+
+    // Obtenir le nom du fichier texte à importer
+    printf("\nGive the complete name of the text file to import: ");
+    pret = fgets(file_import, MAXPATHLEN, stdin);
+
+    if (pret == NULL) {
+        fprintf(stderr, "\nImpossible do read standard input stream\n");
+        exit(1);
+    }
+
+    * index(file_import, '\n') = '\0'; // Supprime le '\n'
+
+    // Ouverture de l'import en mode lecture
+    pf = fopen(file_import, "r");
+
+    if (pf) {
+        Entry * pentry;
+        int nbEntries = 0;
+        char * data;
+
+        do {
+            pentry = malloc(sizeof *pentry); // Allocation de l'entrée
+
+            // Lecture de "information" et suppression de '\n'
+            data = fgets((char *)(pentry->information), MAX_SIZE, pf);
+            if (data) * index((char *)(pentry->information), '\n') = '\0';
+
+            if (data) {
+                 // Lecture de "secret" et suppression de '\n'
+                data = fgets((char *)(pentry->secret), MAX_SIZE, pf);
+                if (data) * index((char *)(pentry->secret), '\n') = '\0';
+
+                if (data) {
+                    // Chiffre l'entrée en mémoire
+                    cypher_data(key, pentry);
+
+                    // Ajout de l'entrée dans la liste
+                    list = addAtLast_DLList(list, pentry);
+
+                    nbEntries++;
+                }
+            }
+
+        } while(data);
+
+        printf("\n%d entries imported from %s\n", nbEntries, file_import);
+        fclose(pf);
+
+    } else {
+        fprintf(stderr, "\nImpossible do open the importation file (%s)\n", file_import);
+        exit(1);
+    }
+
+    return list;
+}
+
 int main(int argc, char* argv[]) {
     char command; // La commande en cours
 
@@ -484,7 +817,8 @@ int main(int argc, char* argv[]) {
 
     DLList list = NULL; // La liste contenant les données chiffrées
 
-    const char file_name[] = "./yatpama.data"; // Le nom et chemin du fichier 
+    const char file_name[] = "./yatpama.data"; // Le nom et chemin du fichier
+    const char file_export[] = "./yatpama_export.txt"; // Nom du fichier d'exportation
 
     do {
         command = prompt();
@@ -492,7 +826,7 @@ int main(int argc, char* argv[]) {
             case 'k':
                 printf("\nEnter password\n");
                 if (!has_key) {
-                    do_command_key(key); // Saisie le mdp et génère la clé
+                    do_command_key(argv[0], key); // Saisie le mdp et génère la clé
                     list = load_data(key, file_name); // Charge et contrôle les données
                     int nbEntries = size_DLList(list);
                     if (nbEntries) printf("\nEntries found in a local data file: %i", nbEntries);
@@ -518,7 +852,7 @@ int main(int argc, char* argv[]) {
                 printf("\nAdd a new secret information\n");
                 if (has_key) {
                     list = do_command_add(key, list);
-                    save_data(list, file_name);
+                    save_data(list, file_name, key);
                 }
                 else
                     printf("...but we don't have password!\n");
@@ -532,7 +866,26 @@ int main(int argc, char* argv[]) {
                     int nbEntries = size_DLList(list);
                     list = do_command_delete(key, list);
                     if (size_DLList(list) == nbEntries - 1)
-                        save_data(list, file_name);
+                        save_data(list, file_name, key);
+                }
+                else
+                    printf("...but we don't have password!\n");
+                break;
+            case 'e':
+                printf("\nExport entries\n");
+                if (has_key) {
+                    do_command_export(key, list, file_export);
+                }
+                else
+                    printf("...but we don't have password!\n");
+                break;
+            case 'i':
+                printf("\nImport entries\n");
+                if (has_key) {
+                    int nbEntries = size_DLList(list);
+                    list = do_command_import(key, list);
+                    if (size_DLList(list) != nbEntries)
+                        save_data(list, file_name, key);
                 }
                 else
                     printf("...but we don't have password!\n");